requirements.txt
streamlit>=1.30.0
streamlit-folium
streamlit-echarts
streamlit-code-editor
folium
openai
python-dotenv

streamlit-ace
streamlit-lightweight-charts
streamlit-option-menu
